fixes for color property dssr xxxxx so that "n/a" is gray
Former-commit-id: 06b6fafaca8f29574eee6f3301da6561f56fcb13

+++ Jmol/src/org/jmol/api/JmolAnnotationParser.java
@@ -22,7 +22,7 @@ public interface JmolAnnotationParser {
 
   String getHBonds(ModelSet ms, int modelIndex, Lst<Bond> vHBonds, boolean doReport);
 
-  void getAtomicDSSRData(ModelSet ms, int modelIndex, int[] dssrData,
+  void getAtomicDSSRData(ModelSet ms, int modelIndex, float[] dssrData,
                          String dataType);
 
   String calculateDSSRStructure(Viewer vwr, BS bsAtoms);
