+++ Jmol/src/org/jmol/modelsetbio/BioModel.java
@@ -438,7 +438,7 @@ public final class BioModel extends Model {
     getRasmolHydrogenBonds(bs, bs, null, false, Integer.MAX_VALUE, false, null, dsspVersion);
   }
 
-  public void getAtomicDSSRData(int[] dssrData, String dataType) {
+  public void getAtomicDSSRData(float[] dssrData, String dataType) {
     if (auxiliaryInfo.containsKey("dssr"))
       vwr.getAnnotationParser(true).getAtomicDSSRData(ms, modelIndex, dssrData, dataType);
   }
